+++ lib/async_redux.dart
@@ -5,6 +5,7 @@ export 'src/navigate_action.dart';
 export 'src/persistor.dart';
 export 'src/store.dart';
 export 'src/store_tester.dart';
+export 'src/mock_store.dart';
 export 'src/user_exception.dart';
 export 'src/wait.dart';
 export 'src/wait_action.dart';